Announce build failure and recovery to configured channels
Note that CircleCI will hit the webhook for each of the jobs in
a workflow.  CircleCI also does not give a status for the workflow as
a whole.

We check the "recovery" announcements on a job-by-job process

If the hubot has a persistent brain, this will always be announced.
However if the recovery is the first web-hook delivery after the reboot
of the hubot instance, the recovery announcement will not be sent

Signed-off-by: Edward Evans <edward.joshua.evans@gmail.com>